nothing

+++ server/controllers/articleController.js
@@ -16,9 +16,7 @@ const createAnArticle = asyncMiddleware(async (req, res, next) => {
   const { title, preview, content, topics } = req.body;
 
   const filename = req.file?.filename;
-  if (!filename) {
-    throw new ErrorResponse(422, "article image is required");
-  }
+  if (!filename) throw new ErrorResponse(422, "article image is required");
 
   const slug = toSlug(title) + "-" + Date.now();
 
@@ -34,9 +32,7 @@ const createAnArticle = asyncMiddleware(async (req, res, next) => {
 
   await article.save();
 
-  res.status(201).json({
-    success: true,
-  });
+  res.status(201).json({ success: true });
 });
 
 // ==================== update article ==================== //
@@ -51,17 +47,11 @@ const updateMyArticle = asyncMiddleware(async (req, res, next) => {
   const oldArticle = await Article.findOne({ slug })
     .lean()
     .select("_id slug img");
-  if (!oldArticle) {
-    throw new ErrorResponse(404, "article not found");
-  }
+  if (!oldArticle) throw new ErrorResponse(404, "article not found");
 
-  let updatedSlug;
-
-  if (title) {
-    updatedSlug = toSlug(title) + "-" + Date.now();
-  } else {
-    updatedSlug = oldArticle.slug;
-  }
+  const updatedSlug = title
+    ? toSlug(title) + "-" + Date.now()
+    : oldArticle.slug;
 
   await Article.findOneAndUpdate(
     { author: me._id, slug },
@@ -69,13 +59,9 @@ const updateMyArticle = asyncMiddleware(async (req, res, next) => {
     { new: true }
   );
 
-  if (filename) {
-    removeFile(oldArticle.img);
-  }
+  if (filename) removeFile(oldArticle.img);
 
-  res.status(200).json({
-    success: true,
-  });
+  res.status(200).json({ success: true });
 });
 
 // ==================== delete article ==================== //
@@ -84,17 +70,15 @@ const deleteMyArticle = asyncMiddleware(async (req, res, next) => {
   const { me } = req;
   const { slug } = req.params;
 
-  const article = await Article.findOneAndDelete({
-    author: me._id,
-    slug,
-  });
-  if (article) {
-    removeFile(article.img);
-  }
+  const article = await Article.findOneAndDelete({ author: me._id, slug });
 
-  res.status(200).json({
-    success: true,
-  });
+  await Comment.deleteMany({ article: article._id });
+
+  await Like.deleteMany({ article: article._id });
+
+  if (article) removeFile(article.img);
+
+  res.status(200).json({ success: true });
 });
 
 // ==================== get an article ==================== //
@@ -106,125 +90,145 @@ const getAnArticle = asyncMiddleware(async (req, res, next) => {
   const article = await Article.findOne({ slug, status: "approved" })
     .lean()
     .select("-status")
-    .populate({ path: "author", select: "avatar fullname username" })
-    .populate({ path: "topics", select: "name slug" });
+    .populate({ path: "topics", select: "name slug" })
+    .populate({ path: "author", select: "avatar fullname username" });
 
-  if (!article) {
-    throw new ErrorResponse(404, "article not found");
-  }
+  if (!article) throw new ErrorResponse(404, "article not found");
 
   article.img = addUrlToImg(article.img);
 
   article.author.avatar = addUrlToImg(article.author.avatar);
 
-  const commentCount = await Comment.countDocuments({ article: article._id });
-
-  const likes = await Like.find({ article: article._id })
-    .lean()
-    .distinct("user");
-
-  const result = { article, likeCount: likes.length, commentCount };
+  const result = { ...article };
 
   if (me) {
+    result.isLiked = !!(await Like.exists({
+      user: me._id,
+      article: article._id,
+    }));
     result.isMyArticle = me._id.toString() === article.author._id.toString();
     if (!result.isMyArticle) {
-      result.authorFollowing = !!(await FollowUser.exists({
+      result.authorFollowed = !!(await FollowUser.exists({
         follower: me._id,
         following: article.author._id,
       }));
     }
-
-    result.isLiked = likes.includes(me._id);
   }
 
-  res.status(200).json({
-    success: true,
-    data: result,
-  });
+  res.status(200).json({ success: true, data: result });
 });
 
-// ==================== get all articles when login ==================== //
+// ==================== count article likes ==================== //
+
+const countArticleLikes = asyncMiddleware(async (req, res, next) => {
+  const { slug } = req.params;
+
+  const article = await Article.exists({ slug, status: "approved" });
+
+  if (!article) throw new ErrorResponse(404, "article not found");
+
+  const count = await Like.count({ article: article._id });
+
+  res.status(200).json({ success: true, data: count });
+});
+
+// ==================== count article commets ==================== //
+
+const countArticleComments = asyncMiddleware(async (req, res, next) => {
+  const { slug } = req.params;
+
+  const article = await Article.exists({ slug, status: "approved" });
+
+  if (!article) throw new ErrorResponse(404, "article not found");
+
+  const count = await Comment.count({ article: article._id });
+
+  res.status(200).json({ success: true, data: count });
+});
+
+// ==================== get all articles ==================== //
 
 const getAllArticles = asyncMiddleware(async (req, res, next) => {
   const { me } = req;
   const { tag, feed } = req.query;
+  const { skip, limit = 15 } = req.query;
 
-  const findQuery = { status: "approved" };
+  const query = { status: "approved" };
+
+  if (skip) query._id = { $lt: skip };
 
   if (tag) {
     const topic = await Topic.exists({ slug: tag });
-    if (!topic) {
-      throw new ErrorResponse(404, "topic tag not found");
-    }
-    findQuery.topics = topic._id;
+    if (!topic) throw new ErrorResponse(404, "topic tag not found");
+    query.topics = topic._id;
   }
 
   if (feed) {
-    const myFollowing = await FollowUser.find({
-      follower: me._id,
-    })
+    const myFollowing = await FollowUser.find({ follower: me._id })
       .lean()
       .distinct("following");
-    findQuery.author = { $in: myFollowing };
+    query.author = { $in: myFollowing };
   }
 
-  const articles = await Article.find(findQuery)
+  const articles = await Article.find(query)
     .lean()
+    .limit(limit)
     .select("-status -content")
+    .populate({ path: "author", select: "avatar fullname username" })
     .populate({ path: "topics", options: { limit: 1 }, select: "name slug" })
-    .sort({ createdAt: -1 })
-    .populate({ path: "author", select: "avatar fullname username" });
+    .sort({ createdAt: -1 });
 
   articles.forEach((article) => {
     article.author.avatar = addUrlToImg(article.author.avatar);
     article.img = addUrlToImg(article.img);
   });
 
-  res.status(200).json({
-    success: true,
-    data: articles,
-  });
+  const skipID = articles.length > 0 ? articles[articles.length - 1]._id : null;
+
+  res.status(200).json({ success: true, data: articles, skipID });
 });
 
 // ==================== search topics for create article ==================== //
 
 const searchTopics = asyncMiddleware(async (req, res, next) => {
   const { search } = req.body;
+  const { skip, limit = 15 } = req.query;
 
   let topics = [];
 
   if (search) {
-    topics = await Topic.find({ name: new RegExp(search, "i") }).lean();
+    const query = { name: { $regex: search, $options: "i" } };
+
+    if (skip) query._id = { $gt: skip };
+
+    topics = await Topic.find(query).lean().limit(limit).sort({ _id: 1 });
   }
 
-  res.status(200).json({
-    success: true,
-    data: topics,
-  });
+  const skipID = topics.length > 0 ? topics[topics.length - 1]._id : null;
+
+  res.status(200).json({ success: true, data: topics, skipID });
 });
 
 // ==================== search articles ==================== //
 
 const searchArticles = asyncMiddleware(async (req, res, next) => {
   const { search } = req.body;
-  const { page = 1, limit = 10 } = req.query;
-
-  const skip = (page - 1) * limit;
+  const { skip, limit = 15 } = req.query;
 
   let articles = [];
 
   if (search) {
-    articles = await Article.find({
-      title: new RegExp(search, "i"),
-      status: "approved",
-    })
+    const query = { $text: { $search: search }, status: "approved" };
+
+    if (skip) query._id = { $lt: skip };
+
+    articles = await Article.find(query)
       .lean()
-      .skip(skip)
       .limit(limit)
       .select("-status -content")
       .populate({ path: "topics", options: { limit: 1 }, select: "name slug" })
-      .sort({ createdAt: -1 })
-      .populate({ path: "author", select: "avatar fullname username" });
+      .populate({ path: "author", select: "avatar fullname username" })
+      .sort({ createdAt: -1 });
 
     articles.forEach((article) => {
       article.author.avatar = addUrlToImg(article.author.avatar);
@@ -232,16 +236,17 @@ const searchArticles = asyncMiddleware(async (req, res, next) => {
     });
   }
 
-  res.status(200).json({
-    success: true,
-    data: articles,
-  });
+  const skipID = articles.length > 0 ? articles[articles.length - 1]._id : null;
+
+  res.status(200).json({ success: true, data: articles, skipID });
 });
 
 module.exports = {
   createAnArticle,
   updateMyArticle,
   getAnArticle,
+  countArticleLikes,
+  countArticleComments,
   deleteMyArticle,
   getAllArticles,
   searchTopics,
